JAVA-21486 Split or move selenium-junit-testng module (moved-4)
