+++ testing-modules/pom.xml
@@ -39,8 +39,8 @@
         <module>powermock</module>
         <module>rest-assured</module>
         <module>rest-testing</module>
-        <module>selenium-junit-testng</module>
-        <module>selenium-webdriver</module>
+        <module>selenium</module>
+        <module>selenium-2</module>
         <module>spring-mockito</module>
         <module>spring-testing-2</module>
         <module>spring-testing</module>
