+++ testing-modules/selenium-2/README.md
@@ -1,0 +1,10 @@
+### Relevant Articles:
+- [Running Selenium Scripts with JMeter](https://www.baeldung.com/selenium-jmeter)
+- [Fixing Selenium WebDriver Executable Path Error](https://www.baeldung.com/java-selenium-webdriver-path-error)
+- [Implicit Wait vs Explicit Wait in Selenium Webdriver](https://www.baeldung.com/selenium-implicit-explicit-wait)
+
+#### Notes:
+- to run the live tests for the article *Fixing Selenium WebDriver Executable Path Error*, follow the manual setup described
+  [Fixing Selenium WebDriver Executable Path Error](https://www.baeldung.com/java-selenium-webdriver-path-error#manual-setup); download the 3
+  drivers mentioned and place them in the src/test/resources directory 
+
