+++ testing-modules/selenium-2/pom.xml
@@ -3,9 +3,9 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>selenium-webdriver</artifactId>
+    <artifactId>selenium-2</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>selenium-webdriver</name>
+    <name>selenium-2</name>
 
     <parent>
         <groupId>com.baeldung</groupId>
