+++ testing-modules/selenium-2/src/test/java/com/baeldung/selenium/setup/AutomatedSetupLiveTest.java
