--- testing-modules/selenium-junit-testng/README.md
@@ -1,18 +1,0 @@
-### Relevant Articles:
-
-- [Guide to Selenium with JUnit / TestNG](http://www.baeldung.com/java-selenium-with-junit-and-testng)
-- [Testing with Selenium/WebDriver and the Page Object Pattern](http://www.baeldung.com/selenium-webdriver-page-object)
-- [Using Cookies With Selenium WebDriver in Java](https://www.baeldung.com/java-selenium-webdriver-cookies)
-- [Clicking Elements in Selenium using JavaScript](https://www.baeldung.com/java-selenium-javascript)
-- [Taking Screenshots With Selenium WebDriver](https://www.baeldung.com/java-selenium-screenshots)
-- [Running Selenium Scripts with JMeter](https://www.baeldung.com/selenium-jmeter)
-- [Fixing Selenium WebDriver Executable Path Error](https://www.baeldung.com/java-selenium-webdriver-path-error)
-- [Handle Browser Tabs With Selenium](https://www.baeldung.com/java-handle-browser-tabs-selenium)
-- [Implicit Wait vs Explicit Wait in Selenium Webdriver](https://www.baeldung.com/selenium-implicit-explicit-wait)
-- [StaleElementReferenceException in Selenium](https://www.baeldung.com/selenium-staleelementreferenceexception)
-- [Retrieve the Value of an HTML Input in Selenium WebDriver](https://www.baeldung.com/java-selenium-html-input-value)
-- [Opening a New Tab Using Selenium WebDriver in Java](https://www.baeldung.com/java-selenium-open-new-tab)
-#### Notes:
-- to run the live tests for the article *Fixing Selenium WebDriver Executable Path Error*, follow the manual setup described 
-[Fixing Selenium WebDriver Executable Path Error](https://www.baeldung.com/java-selenium-webdriver-path-error#manual-setup); download the 3
-drivers mentioned and place them in the src/test/resources directory
