--- testing-modules/selenium-webdriver/README.md
@@ -1,2 +1,0 @@
-### Relevant Articles:
-- [Uploading File Using Selenium Webdriver in Java](https://www.baeldung.com/java-selenium-upload-file)
