+++ testing-modules/selenium/README.md
@@ -1,0 +1,17 @@
+### Relevant Articles:
+
+- [Guide to Selenium with JUnit / TestNG](http://www.baeldung.com/java-selenium-with-junit-and-testng)
+- [Clicking Elements in Selenium using JavaScript](https://www.baeldung.com/java-selenium-javascript)
+- [Handle Browser Tabs With Selenium](https://www.baeldung.com/java-handle-browser-tabs-selenium)
+- [Opening a New Tab Using Selenium WebDriver in Java](https://www.baeldung.com/java-selenium-open-new-tab)
+- [Retrieve the Value of an HTML Input in Selenium WebDriver](https://www.baeldung.com/java-selenium-html-input-value)
+- [Testing with Selenium/WebDriver and the Page Object Pattern](http://www.baeldung.com/selenium-webdriver-page-object)
+- [Using Cookies With Selenium WebDriver in Java](https://www.baeldung.com/java-selenium-webdriver-cookies)
+- [Taking Screenshots With Selenium WebDriver](https://www.baeldung.com/java-selenium-screenshots)
+- [Uploading File Using Selenium Webdriver in Java](https://www.baeldung.com/java-selenium-upload-file)
+- [StaleElementReferenceException in Selenium](https://www.baeldung.com/selenium-staleelementreferenceexception)
+
+#### Notes:
+- to run the live tests, follow the manual setup described 
+[Fixing Selenium WebDriver Executable Path Error](https://www.baeldung.com/java-selenium-webdriver-path-error#manual-setup); download the 3
+drivers mentioned and place them in the src/test/resources directory
