+++ testing-modules/selenium/pom.xml
@@ -3,9 +3,9 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>selenium-junit-testng</artifactId>
+    <artifactId>selenium</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>selenium-junit-testng</name>
+    <name>selenium</name>
 
     <parent>
         <groupId>com.baeldung</groupId>
