+++ testing-modules/selenium/src/test/java/com/baeldung/selenium/tabs/SeleniumOpenNewTabLiveTest.java
